research_assistant/evaluation/requirements.txt
ragas
datasets
pandas
requests
python-dotenv
langchain
langchain-groq
langchain-huggingface
